requirements.txt

# Flask Packages  
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
# Flask-PageDown==0.4.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1

# requests Packages
requests==2.28.2
requests-oauthlib==1.3.1

Werkzeug==2.2.3
WTForms==2.3.3
# Google Packages
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# Database Connections
alembic==1.9.4
psycopg2==2.9.5 
pymongo==4.3.3

# Environments
python-dotenv==1.0.0


